requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
contourpy==1.3.3
cycler==0.12.1
fastapi==0.136.0
flatbuffers==25.12.19
fonttools==4.62.1
h11==0.16.0
idna==3.11
kiwisolver==1.5.0
matplotlib==3.10.8
mpmath==1.3.0
numpy==2.4.4
onnxruntime==1.24.4
opencv-python==4.13.0.92
packaging==26.1
pillow==12.2.0
protobuf==7.34.1
psycopg2-binary==2.9.11
pydantic==2.13.2
pydantic_core==2.46.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
requests==2.33.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
sympy==1.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
